trekk ut felles disply text
