+++ apps/innbyggers-flate/package.json
@@ -23,6 +23,7 @@
     "@navikt/nav-dekoratoren-moduler": "^2.1.6",
     "dayjs": "^1.11.10",
     "deltaker-flate-model": "workspace:*",
+    "deltaker-flate-utils": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.22.1",
