+++ apps/nav-veileders-flate/src/components/pamelding/PameldingFormButtons.tsx
@@ -19,6 +19,7 @@ import {
   useModiaLink
 } from '../../hooks/useModiaLink.ts'
 import { ErrorPage } from '../../pages/ErrorPage.tsx'
+import { DeltakerStatusType } from 'deltaker-flate-model'
 
 interface Props {
   pamelding: PameldingResponse
