+++ packages/utils/.eslintrc.cjs
@@ -1,0 +1,5 @@
+/** @type {import("eslint").Linter.Config} */
+module.exports = {
+  root: true,
+  extends: ['@amt-deltaker-flate/eslint-config/library.js']
+}
