+++ packages/utils/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "deltaker-flate-utils",
+  "version": "0.0.1",
+  "exports": {
+    ".": "./src/utils.ts",
+    "./displayText": "./src/displayText.ts"
+  },
+  "dependencies": {
+    "deltaker-flate-model": "workspace:*",
+    "typescript": "^5.4.4"
+  },
+  "devDependencies": {
+    "@amt-deltaker-flate/eslint-config": "workspace:*"
+  }
+}
