+++ packages/utils/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "@amt-deltaker-flate/typescript-config/base.json",
+  "compilerOptions": {
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ]
+  },
+  "include": ["src"],
+  "exclude": ["node_modules"],
+  "references": [
+    {
+      "path": "./tsconfig.node.json"
+    }
+  ]
+}
